Build activity dictionary from csv
Build an activity dictionary from the activity .csv file that is
loaded.  Also dynamically construct the map layers and keep a
reference to the vector sources in the activity dictionary.  The
key is the activity-id field that is the first field of the .csv
file.  Color-code activities in the list, and display activity
info when the activity from the list is clicked on.