Make activity selection behavior consistent
Remove old "highlighted" style because it is not used anymore.
Selected activities get their style from the default Select
interaction, for now.

Allow multiple activities to be selected on the Activity List
column on the right.  Clicking on the Activity List now toggles
the activity to be in the "selected" list or not.

Clear the Activity List element backgrounds back to transparent
when those activities are no longer selected.